Consolidate release job
Move the homebrew release job so that it's part of the same file with
the other release steps, so we can specify a dependency on the first
job. This fixes the issue where the Homebrew tap update job wasn't being
triggered properly.

In order to work with the automated homebrew action, the package name of
the macOS build of diffsitter has to contain the string "macos", so the
package has been renamed to use "macos" rather than "darwin"