Better diff abstractions, use hunks, better logs (#37)
This commit reworks the abstractions used in the codebase as they
pertain to the actual diff (such as hunks and lines). They are now in
their own module and provide a very convenient interface for building
them up in linear time while we recreate the path from the minimum edit
distance data.

We also create the hunks directly from the predecessor data from the min
distance algorithm, rather than recreating a linear vector of edits from
both documents and grouping the hunks together as a post-process
operation. The new approach is significantly more efficient.

We also introduce "hunks", which is just another term for saying that
edits on adjacent lines will be displayed together because it's
conventional and easier to look at.

Finally, we introduce more granular debug log output, as well as
execution time logging for the parsing methods, the recreate path
method, and the implementation of the minimum edit distance algorithm.