Use a 2D vector instead of hashmap for backtracking
Because we know all of the possible predecessors, we can use a 2D vector
for a significant speedup when computing the minimum edit distance,
since performance profiling shows that the current bottleneck is
computing hashes for indices.

Performance profilign with these changes shows that the bottleneck is
now allocating memory, which is likely as good as it gets.