Simplify config file logic (#86)
Simplify the config file handling, mostly by offloading it to the `xdg`
crate for macOS and linux (or really, anything that's not windows), and
using `directories-next` for Windows, to use the proper path.

This also cleans up the code a bit because we're not manually trying to
deal with different environment variables and whatnot, which is good
because I was doing it incorrectly before anyways.

This commit also moves the different dependencies into the appropriate
conditional compilation targets, since some crates are only used with
certain platforms.