Use a proper buffered writer (#101)
Use the standard library's `BufWriter` to wrap the terminal object so
that the buffer is flushed automatically based on capacity (in this
case, the current default is 8kb). This is more performant than flushing
on every hunk, for example, which incurs a worse performance penalty if
we have many small hunks.

Flushing based on size is a better heuristic and the standard library
`BufWriter` allows us to do this automatically.