Support dynamically linking grammars (#177)
* Support dynamically linking grammars

Add a build flag that allows toggling support for dynamic and static
linking of grammars and add tests for dynamic grammar loading.

* Update CI/CD pipelines

Update the github actions so we specify each feature for debug builds,
tests, and release builds. We can't enable dynamic libraries when
testing release builds because the libraries won't be found on the
system, but we still want to make sure the code for the libraries
builds.

* Fix typo in library file name

* Add extensions to shared library filenames

`libloading` doesn't supply the extension for you so we need to add the
extension based on the platform.

* Add compilation flag for static grammars

Add a compilation flag for static grammars so the binary doesn't attempt
to compile any of the grammars unless the compilation flag is selected.

* Add config for grammars

* Add a proper config for grammars
    * Add configs to set the dynamic library paths for different
      languages
    * Refactor extension overrides into this config (breaking change for
      the configs)
    * Update sample config to reflect this change
* Disable dynamic grammar test by default
* Derive the `default` trait for configs instead of using a manual
  implementation

* Fix target_os string for MacOS

* Add documentation for using dylibs

* Conditionally include codegen file

Don't include the codegen file if the static grammar flag hasn't been
set, because the file won't exist.

* Address clippy lint

Ignore a clippy lint about unused arguments since it's valid with
certain build flags.

* fix compilation errors with build flags

Move everything behind the proper feature gates so each combination of
build flags compiles correctly and address clippy warnings.

* Remove commented line

Remove an unused line

* Fix segfault with dynamic grammars

Leak the symbol table pointer so there aren't any segfaults when working
with the gramamr throughout the program.

* Fix typo in sample config

`c` should be `cpp`

* Test dynamic-grammar-libs in the CI

Enable the feature so tests are run in the CI

* Remove unneeded lifetime parameter

This addresses a clippy lint