feat: Better ordering for diff hunks (#323)
This updates some the logic for aggregating hunks so we can preserve
better ordering around diff hunks.

Before, we would compute the edit script and then split the entries into
their old/new respectively, and then coalesce them into hunks. When
displaying these hunks, we would maintain a pointer to the current
position of the hunks in each document, and then display whichever hunk
had a lower starting line.

This was suboptimal because we ended up losing location information from
the edit script.

We have updated the logic for coalescing hunks to better respect the
original order of the edit script. Now we iterate through the edit
script in order, without doing any preprocessing or splitting based on
the document source. We will aggregate edit entries that are in adjacent
lines, but we are going in the order of the original edit script.

This also cleans up some old unused code in the diff module and adds
some performance improvements by doing unchecked vector indexing in
scenarios where we already check the indices.