feat(output): Implement modular rendering (#446)
Add a generic rendering interface for diffs. This defines a trait
implementation that the rest of the program can use instead of being
tied to a specific implementation.

To enable this we created a rendering trait, which serves as the
interface. We can trivially add new renderers by creating modules that
implement the `Renderer` trait and add them to the enum.

We also add the ability to define "rendering modes" in the config, which
is an arbitrary mapping of tags to a set of rendering settings. Users
can toggle different rendering modes through the command line and
specify the default one to use in the config file.