feat(rendering): Add JSON rendering mode (#519)
This implements a new JSON output mode for diffsitter. We have added
serialization support through the serde macros to the structs passed to
the render interface to make it easier to convert data to JSON.

This is a first pass and we will likely end up adding more information
to the JSON output.