ci: Update Github actions to use more recent APIs (#617)
* ci: Update Github actions to use more recent APIs

This uses cargo nextest and dtolnay's rust toolchain action. The old
`actions-rs` Github actions have been deprecated and unmaintained.

* fix: Remove redundant statement

We already pull in the rust toolchain above and it's invalid to have a
`uses` and `runs` statement in the same block.

* Update uses directive

Previous thing was malformed

* Add components, env var

* Remove redundant env vars

Didn't realize I'd already set it above

* Use talk-e's steps, simplify flow

- We don't need to check the build and also tests, because you need to
  build the binary to test anyways
- Use talki-e's cross compilation steps

* Add aarch64 windows target, fix arm target

* Remove unused parameter

We don't use it anymore

* Install nextest in CI

* Don't use cross toolchain with non linux targets

Turns out it's not supported

* Try to fix conditions

* Put conditional in braces

* Specify toolchain

* Only run doctests on one target

We can't run doctests in the cross toolchain

* Move conditional statement

* disable arm7

Seems like it's not recognized as a valid target

* Publish junit report, enable junit in profile

Add a nextest profile that outputs a junit file, add a junit action and
use the ci profile with nextest

* Use target name

So junit reports don't all have the same name which can be a bit
confusing.

* Break a test

Just for fun to test out the junit annotation action.

* Revert "Break a test"

This reverts commit 9d80a1c3ab171569a31a33376f162d09a173b97f.

* Remove commented out code

* fix: Remove windows aarch64 target

* Install cross toolchains correctly

* Add more release targets

* Fix target triplets

* Add more targets

* Set darwin target

* Load rust cache before other steps

* Use talk-e's upload action

* Add checksum

* Add asset

* Use install action for cargo deb

* Fix install action

* Remove unnecessary steps

* Split up install step

* Fix install action version

* Use production profile

* Don't upload git-diffsitter

It seems like the action tries to upload it for every target in the
matrix

* Build the binary before running cargo deb

* set target directory

* Remove unneeded line

* fix env var declaration

* Add env var for cargo terminal colors

* Install gh cli

* Don't install CLI action

* Put command into 1 line

* Specify target in command

* Remove redundant step

Don't need to build the binary twice

* Use newer version of deb

* Remove homebrew step

Will split this to another action

* Build the production binary

Apparently for cross tools it's recommended to build separately.

* Copy from the correct target dir