ci: Fix upload rust binary action for nightly (#637)
* ci: Fix upload rust binary action for nightly

The action requires the event to be triggered by a tag, which does not
happen for nightlies. In this case we need to manually specify the ref.

* ci: Include the git-diffsitter wrapper

These used to be in the releases before I switched to the new action.

* ci: Use the prod profile in CD

The upload rust action now allows you to specify a specific profile.