fix: Check loaded language ABI
Check that the tree-sitter ABI version reported by the loaded language
is within bounds for the tree-sitter client library that's bundled with
diffsitter. This should give more helpful errors when a user has a
grammar that's built against a tree-sitter version that's not in range.

Previously the program would just segfault which does not make it clear
to users that there is a mismatch between the grammar being loaded and
the library bundled in diffsitter.