fix(grammars): Leak language after loading
Leak the shared library from libloading when loading a tree-sitter
`Language`. This should fix a segfault that can occur when trying to
use the dynamic grammar loading feature.

This also introduces a new binary with utilities for diffsitter, which
are mostly helpful for debugging. We decided to split out the utilities
to keep the interface for the main diffsitter binary less bloated.