feat: Specify custom include paths (#863)
We need to update some of the assumptions in the build script about the
structure of the grammar subdirectories. The ocaml tree-sitter repo was
updated and the include directory is no longer under `src/`.

This commit updates the build script so you can optionally specify
include directories. If you don't, the include dir is set to the old
assumed path.