Update treesitter grammars (#867)
* feat: Specify custom include paths

We need to update some of the assumptions in the build script about the
structure of the grammar subdirectories. The ocaml tree-sitter repo was
updated and the include directory is no longer under `src/`.

This commit updates the build script so you can optionally specify
include directories. If you don't, the include dir is set to the old
assumed path.

* chore(grammars): Update grammars

Done with the updater script

* Update test baselines

Node IDs have been updated because of tree sitter grammar updates