test: Less flaky tests, more succinct output (#939)
Update the snapshot tests so it uses a custom function to generate a
string from the hunks instead of using the debug output. In the past,
tests have broken frequently because of tree-sitter grammar updates.
Typically because a grammar definition adds a new type, or changes
slighlty, which affects fields like `type_id` which fails the regression
test. This is just noise and it doesn't matter what the type ID is.

These new functions also provide a much more compact view of the data,
so it's easier to review test diffs if tests do break.