feat: Better config nesting using figment
This simplifies a bunch of the config handling code and adds proper
support for hierarchical inheritance with config values, including
setting options via environment variables.

Before I was relying on the `Default` trait with serde to handle
default config values, but that led to issues because serde will
only fill in values at the type level, and won't merge keys if a type
is partially defined in the deserialization input.

The previous code also implemented manual fallbacks and didn't support multiple
inheritance.